% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_date_inputs}
\alias{check_date_inputs}
\title{Helper function for statsgrokse()}
\usage{
check_date_inputs(from, to, warn = TRUE)
}
\arguments{
\item{from}{first date of timespan to check}

\item{to}{second date of timespan to check}

\item{warn}{should the function warn if it does, e.g. overwrite parameter
inputs to make things work or just do it and not erport}
}
\description{
Helper function for statsgrokse()
}
