% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsgrokse.R
\name{check_page_inputs}
\alias{check_page_inputs}
\title{helperp function for statsgrokse() that checks and corrects page input}
\usage{
check_page_inputs(page)
}
\arguments{
\item{page}{the page titles to be checked}
}
\description{
helperp function for statsgrokse() that checks and corrects page input
}

