% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{expand_ts_v}
\alias{expand_ts_v}
\title{vectorized version of expand_ts()}
\usage{
expand_ts_v(from, to, by = "day")
}
\arguments{
\item{from}{start of timespan}

\item{to}{end of timespan}

\item{by}{in which time unit expansion should take place: \code{"day"},
\code{"month"}, \code{"year"}}
}
\description{
vectorized version of expand_ts()
}

