% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{expand_ts}
\alias{expand_ts}
\title{Helper function for statsgrokse() that transforms a time span into a series
of months}
\usage{
expand_ts(from, to, by = "day")
}
\arguments{
\item{from}{start of timespan}

\item{to}{end of timespan}

\item{by}{in which time unit expansion should take place: \code{"day"},
\code{"month"}, \code{"year"}}
}
\description{
Helper function for statsgrokse() that transforms a time span into a series
of months
}
