% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsgrokse.R
\name{get_data}
\alias{get_data}
\title{function for getting data (download + extraction)}
\usage{
get_data(df)
}
\arguments{
\item{df}{data.frame defining with columns page, from, to, lang}
}
\description{
function for getting data (download + extraction)
}

