% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_urls.R
\name{prepare_urls}
\alias{prepare_urls}
\title{function preparing URLs for downlaod}
\usage{
prepare_urls(df)
}
\arguments{
\item{df}{data.frame defining with columns page, from, to, lang}
}
\description{
function preparing URLs for downlaod
}
