% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsgrokse.R
\name{statsgrokse}
\alias{statsgrokse}
\title{Function for getting access statistics for wikipedia pages from stats.grok.se}
\usage{
statsgrokse(page, from = "2015-01-01", to = "2015-02-01", lang = "en",
  verbose = TRUE, warn = TRUE)
}
\arguments{
\item{page}{The name of the Wikipedia page as to be found in the URL to the
wikipedia article. If e.g. the URL is:
\code{https://en.wikipedia.org/wiki/Peter_Fox_(musician)}, than the page
name equals to \code{Peter_Fox_(musician)}.}

\item{from}{The starting date of the timespan for which access statistics
should be retrieved - note that there is no data prior to 2007-12-01.
Supply some sort of timestamp e.g. of class POSIXlt, POSIXct, Date, or
character. If the option is of type character it should be in the form of
yyyy-mm-dd.}

\item{to}{The last date for which access statistics should be retrieved.
Supply some sort of timestamp e.g. of class POSIXlt, POSIXct, Date, or
character. If the option is of type character it should be in the form of
yyyy-mm-dd.}

\item{lang}{The language shorthand identifying which Wikipedia access
statistics are to be used: e.g. \code{"en"} for the English version found
at https://en.wikipedia.org, \code{"de"} for the German version found at
https://de.wikipedia.org or perhaps \code{"als"} for the Alemannic dialect
found under https://als.wikipedia.org/.}

\item{verbose}{should function report freely on what it is doing or not}

\item{warn}{should the function warn if it does, e.g. overwrite parameter
inputs to make things work or just do it and not erport}
}
\description{
Function for getting access statistics for wikipedia pages from stats.grok.se
}

